Use widget and modal in pathways
