+++ src/components/OverviewTab.js
@@ -9,7 +9,7 @@ import withStyles from '@material-ui/core/styles/withStyles';
 import KnownDrugsWidget from './KnownDrugs';
 import ChemicalProbesWidget from './ChemicalProbes';
 import RelatedTargetsWidget from './RelatedTargets';
-import PathwaysWidget from './PathwaysWidget';
+import PathwaysWidget from './Pathways';
 import ProteinInformationWidget from './ProteinInformation';
 import CancerBiomarkersWidget from './CancerBiomarkersWidget';
